contracts/nfd/requirements.txt
eth-ape[recommended-plugins]==0.8.9
ape-hardhat==0.8.0
ape-solidity==0.8.2
